style : LoginError 를 LoginInfo 파일로 이동 #24
